#1 Rendering fetched client info works
